LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/scripts/underover-parameters-and-embellished-operator-1-expected.txt


FAIL LowerLimitBaselineDropMin assert_approx_equals: munder: under shift expected 30 +/- 1 but got 0
FAIL LowerLimitGapMin assert_approx_equals: munder: under gap expected 110 +/- 1 but got 0
FAIL UpperLimitBaselineRiseMin assert_approx_equals: mover: over shift expected 50 +/- 1 but got 0
FAIL UpperLimitGapMin assert_approx_equals: mover: over shift expected 70 +/- 1 but got 0
∑
∑

∑
∑

∑
∑

∑
∑
